pyproject.toml
[project]
name = "reward-hub"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery==5.4.0",
    "django==5.1.7",
    "django-redis==5.4.0",
    "psycopg[c, pool]",
    "python-decouple==3.8",
    "redis==5.2.1",
    # API
    "django-cors-headers==4.7.0",
    "djangorestframework==3.15.2",
    "drf-spectacular==0.28.0",
    "drf-standardized-errors==0.14.1",
]

[dependency-groups]
dev = [
    { include-group = "dev_utils" },
    { include-group = "debug" },
    { include-group = "code_quality" },
]
dev_utils = [
    "django-browser-reload==1.16.0",
    "watchfiles==1.0.4",
]
debug = [
    "django-debug-toolbar==4.4.6",
    "django-silk==5.3.1",
]
code_quality = [
    "pre-commit-uv==4.1.4",
    "ruff==0.9.2",
]


# ==== Ruff ====

[tool.ruff]
target-version = "py313"
line-length = 80
show-fixes = true
extend-exclude = [
    ".idea",
    ".venv",
    ".vscode",
    ".pytest_cache",
    "**/migrations/*",
]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 80

[tool.ruff.lint]
allowed-confusables = ["’"]
select = [
    "C90",
    "E",
    "F",
    "I",
    "N",
    "NPY",
    "PD",
    "R",
    "RUF",
    "TRY",
    "UP",
    # flake8 rules
    "ANN",
    "ARG",
    "B",
    "BLE",
    "COM",
    "DJ",
    "DTZ",
    "FBT",
    "G",
    "ICN",
    "ISC",
    "LOG",
    "PT",
    "PTH",
    "PYI",
    "RET",
    "Q",
    "S",
    "SIM",
    "T10",
    "T20",
    "TC",
]
ignore = [
    "ANN401",
    "ARG001",
    "ARG002",
    "COM812",
    "RUF012",
]

[tool.ruff.lint.extend-per-file-ignores]
"test_*.py" = [
    "S101",
    "PLR2004",
]

[tool.ruff.lint.isort]
lines-between-types = 1
combine-as-imports = true
section-order = [
    "future",
    "standard-library",
    "third-party",
    "django",
    "first-party",
    "local-folder",
]
known-first-party = ['apps']

[tool.ruff.lint.isort.sections]
"django" = ["django"]
